resilient-microservices/microservices/auth-service/requirements.txt
Flask==2.3.0
requests==2.31.0
PyJWT==2.8.0
